+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+- Allow passing additional classes when render ([#17](https://github.com/avo-hq/class_variants/pull/17))
 
 ## 0.0.7 (2023-12-07)
 - Add support for compound variants ([#8](https://github.com/avo-hq/class_variants/pull/8))